set options[:verbose] and log dependencies
Temporarily turn on verbose logging and display the
maven attributes of the jars this gem depends on.

Example: creating and installing this gem: 

  com.hazelcast.hazelcast-1.8.5-java.gem

will stop at the install step with an error declaring that 
loj4j is missing -- however there are actually two
dependencies missing: loj4j and slf4j.

The attributes for loj4j and slf4j are displayed on the 
console to make it easier to create and install them.

Should probably be using real option parsing.